Create Task1_sudogame.py
Project 1 : Sudoku Solver
The Sudoku puzzle game is played on a 9 x 9 grid. Within the 
rows and columns are 9 boxes which are made up of 3 x 3 space. 
The purpose of the puzzle is to enter a digit number from 1 to 9 
in each row & a 9 × 9 grid column without duplicating any 
numbers between the row, column or box.
Steps to Build a Python Sudoku Solver
1.Installing pygame module.
2. Import Modules- pygame and requests.
3. Setting Pygame and Building a pygame window
4. Making Sudoku Grid
5. Building API
6. Placing the elements on the board
7. Function to check if grid position is empty or not
8. Function for checking if the entered value is valid
9. Solving Sudoku using a Backtracking algorithm.
10. Quitting the game
▪ Backtracking algorithm – solve a Python Sudoku game 
using a backtracking algorithm.
▪ One by one assigning the numbers to the empty cells. 
▪ Check it before assigning that the same number is 
already present in the current row, column and box. 
▪ After that, assign the proper number using a 
recursive function to check whether it leads to the 
solution or not. 
▪ If it doesn’t lead to the solution then assign the next number